pyproject.toml
[project]
name = "alphagenome-ft"
version = "0.1.8"
description = "Finetuning utilities for AlphaGenome with custom heads and parameter freezing"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Alan Murphy", email = "amurphy@cshl.edu"},
    {name = "Masayuki Nagai", email = "nagai@cshl.edu"}
]
keywords = ["alphagenome", "genomics", "deep-learning", "finetuning", "jax"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "aiohttp>=3.0.0",
    "dm-haiku>=0.0.10",
    "jaxtyping>=0.2.0",
    "optax>=0.2.7",
    "orbax-checkpoint>=0.4.0",
    "pandas>=2.0.0",
    "requests>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "aiohttp>=3.0.0",
    "requests>=2.0.0",
    "pybigwig>=0.3.0"
]

[project.urls]
Homepage = "https://github.com/genomicsxai/alphagenome_ft"
Repository = "https://github.com/genomicsxai/alphagenome_ft"
Issues = "https://github.com/genomicsxai/alphagenome_ft/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["alphagenome_ft*"]

[tool.black]
line-length = 100
target-version = ['py311', 'py312', 'py313']

[tool.ruff]
line-length = 100
target-version = "py311"
